Rename values.py to values_ext.py to conform to SP ext framework
