+++ opendbc/car/chrysler/interface.py
@@ -5,7 +5,7 @@ from opendbc.car.chrysler.carstate import CarState
 from opendbc.car.chrysler.radar_interface import RadarInterface
 from opendbc.car.chrysler.values import CAR, RAM_HD, RAM_DT, RAM_CARS, ChryslerFlags, ChryslerSafetyFlags
 from opendbc.car.interfaces import CarInterfaceBase
-from opendbc.sunnypilot.car.chrysler.values import ChryslerFlagsSP
+from opendbc.sunnypilot.car.chrysler.values_ext import ChryslerFlagsSP
 
 
 class CarInterface(CarInterfaceBase):
